+++ python/ray/serve/_private/client.py
@@ -1,4 +1,5 @@
 import asyncio
+import inspect
 import logging
 import random
 import time
@@ -26,6 +27,7 @@ from ray.serve._private.constants import (
 from ray.serve._private.controller import ServeController
 from ray.serve._private.deploy_utils import get_deploy_args
 from ray.serve._private.deployment_info import DeploymentInfo
+from ray.serve._private.http_util import ASGIAppReplicaWrapper
 from ray.serve._private.utils import get_random_string
 from ray.serve.config import HTTPOptions
 from ray.serve.exceptions import RayServeException
@@ -332,6 +334,9 @@ class ServeControllerClient:
 
             name_to_deployment_args_list[app.name] = deployment_args_list
 
+        # Validate applications before sending to controller
+        self._check_ingress_deployments(built_apps)
+
         ray.get(
             self._controller.deploy_applications.remote(name_to_deployment_args_list)
         )
@@ -392,6 +397,29 @@ class ServeControllerClient:
             else:
                 raise TimeoutError(
                     f"Serve application isn't running after {timeout_s}s."
+                )
+
+    def _check_ingress_deployments(
+        self, built_apps: Sequence[BuiltApplication]
+    ) -> None:
+        """Check @serve.ingress of deployments across applications.
+
+        Raises: RayServeException if more than one @serve.ingress
+            is found among deployments in any single application.
+        """
+        for app in built_apps:
+            num_ingress_deployments = 0
+            for deployment in app.deployments:
+                if inspect.isclass(deployment.func_or_class) and issubclass(
+                    deployment.func_or_class, ASGIAppReplicaWrapper
+                ):
+                    num_ingress_deployments += 1
+
+            if num_ingress_deployments > 1:
+                raise RayServeException(
+                    f'Found multiple FastAPI deployments in application "{app.name}".'
+                    "Please only include one deployment with @serve.ingress "
+                    "in your application to avoid this issue."
                 )
 
     @_ensure_connected
